Replace `grep`,`cut` and `sed` commands
Replace `grep`,`cut` and `sed` commands with separate Bash functions. 

These functions does basically the same as `grep`, `cut` and `sed` do, but are a pure Bash alternative.
Because it's pure Bash no sub shells are used and the application is a little bit more efficient.

Because of this change `grep`, `cut` and `sed` are no  longer a required dependency.