Fix bug with password length input validation ReGeX
This commit fixes a bug where the user was unable to select a password length containing a '0'. It accomplishes this by modifying the regular expression to allow for zero, while still ensuring that the input is numeric. Additionally, this commit ensures that the password length is at least one character long, so that we can generate a password without causing 'dd' to complain about a negative count.