Remove raw password file on error
This commit improves the security of BashPass by adding a line to 'the error_out()' function that removes the raw password file if it exists. This ensures that the user does not accidentally leak the password. Although 'rm' is not a built-in Bash command, it is necessary in this case, as there is no other way to achieve this. Additionally, the command is made more secure by using 'rm -rf' to force the removal of the file, and by using the '${variable:?}' syntax to prevent any potential issues with undefined variables.