Remove unnecessary use of sudo/doas
From now on `make` must be ran as a privileged user, such as root.
For BSD you can use `doas`, Linux and macOS users can use `sudo`
to gain such privileges.

This change made the necessity for a separate `install` and `update`
option obsolete, thus it has also been removed.

The pipeline to test the Makefile on Ubuntu has also been altered to
use sudo and remove the `update` test.