Add build pipeline for deb-pkgs
Pipelines regarding the build process of deb-pkgs are now part of the
main BashPass repo, instead of having its own "BasPass-DEB". This
simplifies the release of newer versions, because of the auto-trigger
when a new tag is made, indicating a new release.

Since git(1), wl-clipboard(1) and other clipboard managers are not a
hard dependeny, they will have to be installed separately. This also
makes packageing much easier, since we can ignore xorg and wayland.